+++ crates/editor/src/editor_tests.rs
@@ -24071,119 +24071,6 @@ async fn test_newline_replacement_in_single_line(cx: &mut TestAppContext) {
     });
 }
 
-#[gpui::test]
-async fn test_completion_menu_scroll_aside(cx: &mut TestAppContext) {
-    init_test(cx, |_| {});
-
-    // In order to be able to test the shortcuts that should allow the user to
-    // scroll the aside content of the completions menu, we first need to load
-    // the keymap file, otherwise simulating the keystrokes will not actually
-    // dispatching the necessary action.
-    cx.update(|cx| {
-        let default_key_bindings = settings::KeymapFile::load_asset_allow_partial_failure(
-            "keymaps/default-macos.json",
-            cx,
-        )
-        .expect("Should be able to load deafult MacOS Keymap");
-
-        cx.bind_keys(default_key_bindings);
-    });
-
-    // Setup the completion to be provided by the LSP so we can set the
-    // completion item's documentation with a string long enough to overflow the
-    // assigned vertical space.
-    let mut cx = EditorLspTestContext::new_rust(
-        lsp::ServerCapabilities {
-            completion_provider: Some(lsp::CompletionOptions {
-                trigger_characters: Some(vec![".".to_string()]),
-                ..Default::default()
-            }),
-            ..Default::default()
-        },
-        cx,
-    )
-    .await;
-
-    cx.lsp
-        .set_request_handler::<lsp::request::Completion, _, _>(move |_, _| async move {
-            Ok(Some(lsp::CompletionResponse::Array(vec![
-                lsp::CompletionItem {
-                    label: "Test Item".to_string(),
-                    documentation: Some(lsp::Documentation::String(
-                        "This is some very long documentation content that will be displayed in the aside panel for scrolling.\n".repeat(50)
-                    )),
-                    ..Default::default()
-                },
-            ])))
-        });
-
-    cx.set_state("variableˇ");
-    cx.simulate_keystroke(".");
-    cx.executor().run_until_parked();
-
-    // Verify that the completion menu is visible before attempting to scroll
-    // it's aside content.
-    let mut initial_offset: Pixels = px(0.0);
-
-    cx.update_editor(|editor, _, _| {
-        let binding = editor.context_menu.borrow();
-        let Some(CodeContextMenu::Completions(menu)) = binding.as_ref() else {
-            panic!("Should have completions menu open");
-        };
-
-        assert_eq!(completion_menu_entries(&menu), &["Test Item"]);
-        initial_offset = menu.scroll_handle_aside.offset().y;
-    });
-
-    // The `ctrl-e` shortcut should scroll the completion menu's aside content
-    // down, so the updated offset should be lower than the initial offset.
-    cx.simulate_keystroke("ctrl-e");
-    cx.update_editor(|editor, _, _| {
-        let binding = editor.context_menu.borrow();
-        let Some(CodeContextMenu::Completions(menu)) = binding.as_ref() else {
-            panic!("Should have completions menu open");
-        };
-
-        assert!(menu.scroll_handle_aside.offset().y < initial_offset);
-    });
-
-    // The `ctrl-y` shortcut should do the inverse scrolling as `ctrl-e`, so the
-    // offset should now be the same as the initial offset.
-    cx.simulate_keystroke("ctrl-y");
-    cx.update_editor(|editor, _, _| {
-        let binding = editor.context_menu.borrow();
-        let Some(CodeContextMenu::Completions(menu)) = binding.as_ref() else {
-            panic!("Should have completions menu open");
-        };
-
-        assert_eq!(menu.scroll_handle_aside.offset().y, initial_offset);
-    });
-
-    // The `ctrl-d` shortcut should scroll the completion menu's aside content
-    // down, so the updated offset should be lower than the initial offset.
-    cx.simulate_keystroke("ctrl-d");
-    cx.update_editor(|editor, _, _| {
-        let binding = editor.context_menu.borrow();
-        let Some(CodeContextMenu::Completions(menu)) = binding.as_ref() else {
-            panic!("Should have completions menu open");
-        };
-
-        assert!(menu.scroll_handle_aside.offset().y < initial_offset);
-    });
-
-    // The `ctrl-u` shortcut should do the inverse scrolling as `ctrl-u`, so the
-    // offset should now be the same as the initial offset.
-    cx.simulate_keystroke("ctrl-u");
-    cx.update_editor(|editor, _, _| {
-        let binding = editor.context_menu.borrow();
-        let Some(CodeContextMenu::Completions(menu)) = binding.as_ref() else {
-            panic!("Should have completions menu open");
-        };
-
-        assert_eq!(menu.scroll_handle_aside.offset().y, initial_offset);
-    });
-}
-
 #[track_caller]
 fn extract_color_inlays(editor: &Editor, cx: &App) -> Vec<Rgba> {
     editor
